完成 RBAC 权限功能

+++ backend/controllers/PostController.php
@@ -8,6 +8,7 @@ use common\models\PostSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\web\ForbiddenHttpException;
 
 /**
  * PostController implements the CRUD actions for Post model.
@@ -64,6 +65,10 @@ class PostController extends Controller
      */
     public function actionCreate()
     {
+        if(!Yii::$app->user->can('createPost')) {
+            throw new ForbiddenHttpException('对不起，您没有权限进行该操作！');
+        }
+
         $model = new Post();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -84,6 +89,10 @@ class PostController extends Controller
      */
     public function actionUpdate($id)
     {
+        if(!Yii::$app->user->can('updatePost')) {
+            throw new ForbiddenHttpException('对不起，您没有权限进行该操作！');
+        }
+
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -104,6 +113,10 @@ class PostController extends Controller
      */
     public function actionDelete($id)
     {
+        if(!Yii::$app->user->can('deletePost')) {
+            throw new ForbiddenHttpException('对不起，您没有权限进行该操作！');
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
